requirements.txt
chainlit==2.5.5
python-dotenv==1.1.0
azure-ai-projects==1.0.0b12
azure-ai-inference==1.0.0b6
azure-identity==1.23.1
azure-ai-ml==1.28.1
markitdown[all]==0.1.2
openai==1.82.1
loguru==0.7.3
litellm==1.72.6
asyncpg==0.30.0
azure-storage-blob==12.25.1
aiohttp==3.12.6
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
langchain-azure-ai --only-binary :all:
langgraph
markdownify
tavily-python
fastapi
uvicorn
websockets
pydantic
python-multipart
jinja2
rich_click
azure-mgmt-cognitiveservices
ipython
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0